Renamed ff/diskio => ff_disk_io (naming consistency)
